requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
aiohttp
python-dotenv